fix: make the render-target parity test importable in the server CI job
Main's CI has been red since the Phase 2 parity test landed: it imported
client/src/lib/imageGenBackends.js, whose top-level lucide-react import
cannot resolve in the server job (client packages aren't installed there),
and the unprotected-main auto-merges let the red slip through. The pure
constants + helpers now live in a dependency-free imageGenModes.js leaf
(re-exported flat by imageGenBackends so consumers keep one import site;
namespace-exported in the lib barrel to avoid identifier collisions) and
the parity test imports the leaf.
