+++ server/lib/renderTargets.parity.test.js
@@ -16,10 +16,13 @@
 
 import { describe, it, expect } from 'vitest';
 import { RENDER_TARGETS, RENDER_TARGET_BACKEND_AUTO } from './renderTargets.js';
+// Import the node-safe leaf, NOT imageGenBackends.js — that module imports
+// lucide-react, which is not installed in the server CI job (this exact import
+// broke main's CI when Phase 2 landed pointing at imageGenBackends).
 import {
   RENDER_TARGET_OPTIONS as CLIENT_OPTIONS,
   RENDER_TARGET_BACKEND_AUTO as CLIENT_AUTO,
-} from '../../client/src/lib/imageGenBackends.js';
+} from '../../client/src/lib/imageGenModes.js';
 
 // Targets the Settings UI deliberately does NOT list — a pin nobody's
 // resolver reads would be a control that silently does nothing. Empty since
